Redesigned main menu fixed a bug regarding mui buttons & router links. Restructured components folder into ui and abs(abstraction) components.
