+++ src/components/GameWindow.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import {Box, Typography} from '@material-ui/core';
+import {Box} from '@material-ui/core';
 import {Switch, Route, Redirect} from 'react-router-dom';
 import routes from '../routes';
-import CenteringBox from './CenteringBox';
-import MainMenu from './MainMenu';
+import CenteringBox from './abs/CenteringBox';
+import MainMenu from './ui/MainMenu';
 import Levels from './Levels';
 import GameLayout from '../layouts/GameLayout';
 
@@ -14,7 +14,7 @@ import GameLayout from '../layouts/GameLayout';
 const GameWindow = () => {
   return (
     <CenteringBox id="game-window" height="100vh" width="100vw" bgcolor="coral">
-      <Box width="1080px" height="540px" id="game" bgcolor="lightblue">
+      <Box width="720px" height="540px" id="game" bgcolor="lightblue">
         <GameLayout>
           <Switch>
             <Route exact path="/">
